Add cart and directory for user

+++ Bookstore_OOP/Model/Book.cs
@@ -8,16 +8,16 @@ using System.Threading.Tasks;
 
 namespace Bookstore_OOP.Model
 {
-    public class Book
-    {
-        [Key]
-        [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int Id { get; set; }
-        public string Title { get; set; }
-        public int AuthorID { get; set; }
-        public string Publisher { get; set; }
-        public int Year { get; set; }
-        public string Genre { get; set; }
-        public int Price { get; set; }
-    }
+        public class Book
+        {
+            [Key]
+            [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
+            public int Id { get; set; }
+            public string Title { get; set; }
+            public int AuthorID { get; set; }
+            public string Publisher { get; set; }
+            public int Year { get; set; }
+            public string Genre { get; set; }
+            public decimal Price { get; set; }
+        }
 }
